requirements.txt
# Streamlit and its core dependencies
streamlit>=1.10.0

# Data manipulation and numerical operations
pandas>=1.3.0
numpy>=1.21.0

# Visualization libraries
plotly>=5.0.0
matplotlib>=3.8.0
seaborn>=0.11.0
missingno>=0.5.1
kaleido==0.2.1

# Machine learning libraries
scikit-learn>=1.0.0
scipy>=1.7.0

# LLM integration with LangChain and OpenAI
openai>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.1.0
chromadb>=0.5.5  # Optional, only if needed
tiktoken>=0.4.0  # Tokenizer for LLM input estimation

# Pydantic - specified to avoid conflicts
pydantic>=2.7.4,<3.0.0



# Environment variable management
python-dotenv>=1.0.0

langchain-cohere